+++ src/app/projects/project/graph/components/script/script-editor/script-editor.component.ts
@@ -55,9 +55,12 @@ export class ScriptEditorComponent implements PropertiesDialogChild, AfterViewIn
         }
     }
 
-    resized(): void {
+    resized(width: number, height: number): void {
         if (this.editor) {
-            this.editor.layout()
+            this.editor.layout({
+                width,
+                height
+            })
         }
     }