+++ src/app/properties-dialog/service/properties-dialog-child.ts
@@ -3,5 +3,5 @@ import { PropertiesDialogServiceEvent } from './properties-dialog-service.event'
 export interface PropertiesDialogChild {
     submit(): Promise<void>
     cancel(): Promise<void>
-    resized(): void
+    resized(width: number, height: number): void
 }